[ipq806x] Board specific image flashing in pack.py
A new field has been added to flash.conf 'if_machid', when the field
is set to a machine type integer, the image corresponding to that
section is flashed only if the U-Boot 'machid' is equal to value
specified in the flash.conf.

This has been added to support multiple boards with a single packed
image.

CRs Fixed: 537629
Change-Id: I29293aab6c2b3745eaf56195c94a7e8b1bf29813
Signed-off-by: Vijay Kumar <c_vbagav@codeaurora.org>
Reviewed-by: Varadarajan Narayanan <varada@codeaurora.org>