[ipq806x] Fix incorrect layout switching
The flashing script does some smart layout switching. If we are
already in the 'sbl' layout it will not trigger a switch to the 'sbl'
layout. The conditional flashing of images interferes with this smart
layout switching logic. Hence disable the smart layout switch, which
currently is only an unnecessary optimization.

CRs Fixed: 537629
Change-Id: I882e49174cba8d2c8587af8f897e87ce08ec51ab
Signed-off-by: Vijay Kumar <c_vbagav@codeaurora.org>
Reviewed-by: Kirthik Srinivasan <kirthik@codeaurora.org>
Reviewed-by: Varadarajan Narayanan <varada@codeaurora.org>