ipq806x: Integrate Crash Scope for Wlan Firmware Assert
When the target asserts, the ram from target is read and copied
to a fixed memory location in host ram at 0x40800000.
So allocate a block of memory in host ram.
The crash scope tool will parse the dump from this memory location.
The content in this memory block is copied to a new file wlan_fw.bin.

Change-Id: I4a083e3bd7d4c994fd41fa67baa683f5b7818d62
CRs-Fixed: 592514
Signed-off-by: Kirthik <kirthik@codeaurora.org>
Signed-off-by: Mohammed Riyaz Piyarajohn <mpiyaraj@codeaurora.org>