Rename sha1.h into u-boot-sha1.h
If the host system has an sha1.h installed, then it will be included
at build time, which will result in a build failure as the one in the
u-boot/include folder will not be used.

Therefore, this change renames sha1.h in u-boot-sha1.h; this will make
sure the header doesn't conflict with any header in the system.

Change-Id: I6da36d43abe2ae6b9c3207fbf0f3aaae4059e3b7
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>