common: Prepend variables with "0" rather than " " in imxinfo
When using imxinfo to extract information from a FIT image, the fileaddr
and filesize variables are prepending with spaces to fit an 8 characters
string. These spaces causes various problems when reusing these values
in scripts afterwards, especially with the "sf" command, which doesn't
expect the file size to be specified as "+${filesize}".

In this situation, the space is interpreted as an argument delimiter,
which causes the command to fail.

Change-Id: Idc0777400de9c2f1b2bf8f2f40d8a12529241818
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>