usbh/ehci: Increase timeout for enumeration
The current logic reads the port status just once after usb_hub_power_on and
expects the portstatus and portchange to report the connection status
immediately and correctly.

Few pen drives are not able to report both of them immediately ie. those pens
report the connection change but not the connected state after the first read.
This opportunity once lost is gone for ever because the u-boot, unlike linux or
any other OS, works in polling mode.

This patch modifies the logic to read the port status continuously until the
portstatus and portchange both report a connection change as well as a connected
state or no connection change and no connection. This logic is placed in a
timeout of 10 sec. At the end of it, the pen drive would have either reported a
ONE or a ZERO in bit 1 of portstatus as well as portchange.

It enhances the set of pen drives which can eventually be detected by u-boot

Note: This 10 second timeout is based purely on several experiments done with
the broken pen drives

Backport XHCI specific code from top of tree U-Boot.
The commit is
http://git.denx.de/?p=u-boot.git;a=commit;h=b6d7852cf8b7ced786937d3d080956b0c7be9836
    usbh/ehci: Increase timeout for enumeration

CRs-Fixed :625723

Change-Id: I418cf8cc7e558d43c19721f2f4eb4b393a7463e1
Signed-off-by: Vipin Kumar <vipin.kumar@st.com>
Acked-by: Igor Grinberg <grinberg@compulab.co.il>
Signed-off-by: Sham Muthayyan <smuthayy@codeaurora.org>