ipq806x: bootmbn cmd for booting kernel in mbn format
The kernel image is converted to mbn format by
adding 40 byte header above the uImage header during the build.

The format conversion was done because Code signing tools expects
the image to be in mbn format. mbn format of kernel image is going
to be maintained for both secured boot and non secured boot.

Added a new command do_bootmbn to boot kernel in mbn format.

Change-Id: I611370dc2374f3cf0897c2454db1aa7b65758ad7
Signed-off-by: Ravindranath Thiyagarajan <rthiyaga@codeaurora.org>