ipq806x: Rework the GPIO code to reflect the datasheet content
GPIO code in its current form doesn't actually reflect the datasheet
content. This change is replacing the fields currently known as
"enable" by "oe", and the "dir" (aka direction) by "out".
Interestingly, the "dir" field is actually unused in today's code, which
means that the Pinmux can be configured to assign functions to certain
GPIOs, but it cannot be used to set a GPIO to 0 or 1 explicitly.

This is a problem if there is a necessity to toggle a GPIO at init.
Certain platform use GPIO as reset lines for other components, in which
case forcing GPIOs to a certain value is required as part of the U-boot
init code.

This change should not include any functional change. Additional changes
will be coming in later patches.

Change-Id: If7e4064e3f441961343b23cf4abbe9d8bd770118
Signed-off-by: Mathieu Olivari <mathieu@codeaurora.org>