ipq4019: Support for W25M02GV Winbond SPI NAND flash
W25M02GV consists of two stacked dies each of 128MB.
Only a single die can be active at a time and hence we
require a die select command to be given along with the
die ID (0 or 1)  before read, write and erase
operations can be done in a particular die. The support
for die select has been implemented in the driver.

Change-Id: I803d4f924f43f31c540458261760779acea7dd50
Signed-off-by: G Dhivya <gdhivya@codeaurora.org>