ARM: env_nand: Updated env to use complete partition
nand env size was restricted to 256KB out of
the available 512KB to keep it compatible with
kernel. When a badblock is introduced,
env fails to save due to the restricted range.

With this patch, it can now use upto the end
of partition upon skipping any badblocks
but only saving 256KB.

Change-Id: I27f6932fe537fa4c63d890bfe583c48e45fe156b
Signed-off-by: Gokul Sriram Palanisamy <gokulsri@codeaurora.org>