ipq807x: uboot: spi init: code clean up
This patch involves changes which makes
sure that spi init is done only when the
config CONFIG_QCA_SPI is defined.

Previously, since spi was a sub part of
nand, spi_init was done without checking if
CONFIG_QCA_SPI is enabled.

This change will allow us to disable or
enable spi as per our need in uboot.

Change-Id: Ic75301872d5ac161cd7b863c59c12437aa6f7126
Signed-off-by: speriaka <speriaka@codeaurora.org>