ipq806x: use proper partition for env varibales in eMMC
In eMMC flash, whenever we save the env variables, it is updated in
the wrong offset instead of using the APPSBLENV partition, which leads
to the system hang in the next reboot. This patch addresses the same.

Also, moved the function board_mmc_env_init to common/env.c, since
the same functionality is used across ipq806x/ipq40xx/ipq807x.

Change-Id: I7e85cbb004ab1ca4d7c926359a8d9a1912922971
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>