ipq806x: use proper partition for env varibales in eMMC
In eMMC flash, whenever we save the env variables, it is updated in
the wrong offset instead of using the APPSBLENV partition, which leads
to the system hang in the next reboot. This patch addresses the same.

Also, moved the function board_mmc_env_init to common/env.c, since
the same functionality is used across ipq806x/ipq40xx/ipq807x.

Change-Id: I4c446dad583089fb2515e13267aa7b4724c16fba
Signed-off-by: Kathiravan T <kathirav@codeaurora.org>