ipq40xx: Register SPI nor as mtd device
Registering SPI nor flash as mtd device was based
on 'spi_gpio' entry in dts, whereas it will be done
based on 'spi_nor_available' in this patch as spi_gpio
entry is not required in some of the SPI nor
supported boards.

This patch also enables the spi nand support
in DK04-c5 board.

Change-Id: I2d70a6c108321d0f504daab7bc4e74739499582a