ipq40xx: Fix for USB 3.0 detection
During usb start, all the hub initialization happens,
before scanning the bus for devices. In DK, bus 0 has 
2 NbrPorts and bus 1 has 1 NbrPorts. Since there are
two ports(USB0 and USB1) available, such initialization 
overwrites the NbrPorts of the hub descriptor. Hence it
fails to detect the USB 3.0.

This patch modifies the logic to scan for devices in the 
hub before initializing the next hub.

Change-Id: I1071d7c38a896864fe15d8b07bdb2a84ddcac7bf
Signed-off-by: Sasirekaa Madhesu <smadhesu@codeaurora.org>