ipq40xx: Changed the dumpinfo name for secure board
In non-secured boards the dump save size is set to 
be the size of RAM buffer. Since the name field of
dumpinfo was same for both secured and non-secured 
boards,the dump save size for secured board is also
set to RAM buffer size which causes data abort during
secure crashdump collection. This patch retains the
actual save size of secure board.

Change-Id: Ib26fc4730355a7f0cb131dee61983a15c1825f6f
Signed-off-by: Sasirekaa Madhesu <smadhesu@codeaurora.org>