ipq807x: scm: Make SCM calls dcache aware
U-Boot now has dcache turned on. This requires cache flushing to
be performed before executing smc instrution. There are calls which
are made to TZ with huge buffer (such as kernel image for authentication).
Flushing and invalidating by address for such huge buffer will have
a performance impact, as the size of buffer is vastly bigger than
the cache size itself. Hence performing a flush of the entire cache
as opposed to performing cache flush with a start and end address.

Change-Id: If90a49d240ecf021ac9714025b6de0cce3fb9fc8
Signed-off-by: Manoharan Vijaya Raghavan <mraghava@codeaurora.org>