board: qca: common: Add condition for NOR boot
In case of NOR boot mount the rootfs based on device
name not based on UUID. Check the flash type based on rootfs
offset. In case of NOR plus eMMC the rootfs offset will be 0xBAD0FF5E.
This is bad offset. In NOR only boot rootfs have proper offset.

Change-Id: I7a7ecf0c1d0f2dcec13c71561568d1e16686316b
Signed-off-by: Md Sadre Alam <mdalam@codeaurora.org>