ipq40xx: mmc: Added check for status of sdhci node
Currently, we search for "sdhci" node and continue with
mmc_init if it is present. We never check for status of
the sdhci node. As sdhci node is present in all the
board variants, mmc_init happens in all boards.

So, added a check to see if the status of sdhci node is
"okay" and continue with initialization, else return.

fdtdec_get_is_enabled library function checks only for
string "okay" which is the correct usage for status
variable. So, changed the status string in the
dts files accordingly.

Change-Id: I26ac5d9930b9e91e3d9d2436eff0cf02b8977b2e
Signed-off-by: Balaji Jagadeesan <bjagadee@codeaurora.org>