ipq806x: Initialize complete ethernet subsystem for AP161
AP161 board has 3 GMAC's GMAC0, GMAC1 and GMAC2. Currently uboot
is initializing GMAC1 and GMAC2 only, not GMAC0.
This way when we use upstream GMAC driver (stmmac dwmac-ipq806x.c)
then GMAC1 and GMAC2 are working fine but GMAC0 does not work since
it is not initialized/configured at uboot.

This change adds support to initialize GMAC0 according to it's
corresponding configuration so that all GMAC's can be used.

Change-Id: I6345282cd48cb3a766f33d124f36d4050e27da00
Signed-off-by: Ram Chandra Jangir <rjangir@codeaurora.org>