ipq806x: Send STOP_TRANSMISSION command after DATA_TIMEOUT error
The Device will accept the requested number of data blocks, terminate
the transaction and return to transfer state. Stop command is not
required at the end of this type of multiple blocks write unless
terminated with an error.  This change will send the
STOP_TRANIMISSION command if the command failed.

Change-Id: I9bd419ab8931d80a4a2eeba9f5bd42222257a824
Signed-off-by: Vinoth Gnanasekaran <vgnana@codeaurora.org>