ipq807x: Disable dump collection during warm reset.
This is to provide software framework to retain IMEM’s state
across warm reset. For warm reset crashdump collection should be
disabled.

In uboot we will check for the environment variable dload_warm_reset.The
crashdump magic is reset to 0x10 in SDI path, hence disable the SDI bit
in Uboot  if environment variable is set which will disable crashdump 
collection. 

Change-Id: Ic60b63ead03436689ee04d2e3e7ecdbb275ccfee
Signed-off-by: Nikhil Prakash V <nprakash@codeaurora.org>