drivers: net: ipq9574: Update EDMA Driver
This patch updates the address of the EDMA registers, removes
registers which are not used and then does the required changes
to accomodate the new EDMA Descriptor format.

This patch also does the required changes to accomodate Secondary
Descriptor which is added instead of preheaders in ipq95xx.

We have also updated the number of descriptors per ring to 128
with this patch and have also moved the configs not required for
EMU Platform to !CONFIG_IPQ9574_RUMI.

Change-Id: I6e98c457d76c66f87b74cc0eafa7088d0081f533
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>