drivers: net: ipq9574: Update PPE Driver
This patch updates the VSI settings. The VSI Table size has
now changes from 3 word to 4 word and the same has been
accomodated. Without writing 4 words to that Table, the VSI
writes won't be processed by the H/W.

This patch updates the scheduling, TDM configuration
and the ppe port mux configuration as required for ipq95xx.

This patch also moves the configurations not required
for EMU Platform to !CONFIG_IPQ9574_RUMI.

Change-Id: Id54e40d26e80c36e7a61642d8494c30bbd3ea2a5
Signed-off-by: Selvam Sathappan Periakaruppan <speriaka@codeaurora.org>